Only fetch latest N posts
Previously, all posts and replies were being fetched for the get_posts
handler. This modifies the endpoint so that it will only return the
latest N posts (and their replies), placing an upper bound on the size
of the response.

Since we currently have the following constraints:

* Body <= 512 chars
* Author name <= 36 chars
* Attachment URL <= 256 chars
* Number of posts = 20

The upper bound of our response is 20 * (512 + 36 + A) + B, where
A is the length of the extra JSON data around each post, and B is the
length of the extra JSON data enclosing the list of posts. Technically
there is also some variability around the length of e.g. integer IDs as
well.